services/git_cloning_service/requirements.txt
gitpython==3.1.40
kafka-python==2.0.2
psycopg2-binary==2.9.9